# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "E3Di & Jaan LED Screen IoT Ad-Network Platform API",
    "description": "Comprehensive REST & IoT API documentation for the E3Di Smart Pole, Digital Billboard Network, Edge Gateways, CCTV surveillance, and Emergency SOS infrastructure.",
    "version": "2.0.0",
    "contact": {
      "name": "E3Di Platform Support",
      "url": "https://e3di.org"
    }
  },
  "servers": [
    {
      "url": "https://api.e3di.org",
      "description": "Production LED CPU API Server"
    },
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server (Port 5000)"
    },
    {
      "url": "http://localhost:5001",
      "description": "Alternative Local Server (Port 5001)"
    }
  ],
...
  "paths": {
...
    "/api/edges/upload-segment": {
      "post": {
        "tags": ["IoT Edge Gateway"],
        "summary": "Upload Raw Binary HLS Video Segment",
        "description": "Edge agent gateway uploads binary `.ts` MPEG-2 video transport segments directly from local RTSP CCTV streams with atomic write.",
        "parameters": [
          { "name": "x-camera-id", "in": "header", "required": true, "schema": { "type": "string" } },
          { "name": "x-filename", "in": "header", "required": true, "schema": { "type": "string" } },
          { "name": "x-edge-id", "in": "header", "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "*/*": {
              "schema": { "type": "string", "format": "binary" }
            }
          }
        },
        "responses": {
          "200": { "description": "Segment written successfully" },
          "400": { "description": "Missing camera ID, filename or body" }
        }
      }
    },
    "/api/admin/edges/register": {
      "post": {
        "tags": ["IoT Edge Gateway"],
        "summary": "Register Edge Agent Gateway",
        "description": "Registers a new site-level edge agent gateway responsible for local CCTV RTSP ingest and LED screen monitoring.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["edgeId", "siteId"],
                "properties": {
                  "edgeId": { "type": "string", "example": "EDGE-001" },
                  "siteId": { "type": "string", "example": "SITE-001" },
                  "name": { "type": "string", "example": "Vijayawada MG Road Gateway" },
                  "agentVersion": { "type": "string", "example": "1.0.0" },
                  "localIp": { "type": "string", "example": "192.168.1.107" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Edge agent registered successfully" }
        }
      }
    },
    "/api/edges/{edgeId}/heartbeat": {
      "post": {
        "tags": ["IoT Edge Gateway"],
        "summary": "Edge Agent Heartbeat & Telemetry",
        "description": "Reports edge gateway CPU load, memory utilization, active RTSP transcode streams, and CCTV health.",
        "parameters": [
          { "name": "edgeId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cpuUsage": { "type": "number", "example": 24.5 },
                  "memoryUsage": { "type": "number", "example": 42.1 },
                  "activeStreams": { "type": "integer", "example": 2 },
                  "cameraStatus": { "type": "string", "example": "online" },
                  "uptime": { "type": "number", "example": 86400 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Heartbeat accepted" }
        }
      }
    },
    "/api/admin/edges": {
      "get": {
        "tags": ["IoT Edge Gateway"],
        "summary": "List All Registered Edge Gateways",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of edge agents across all sites",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "edges": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/EdgeAgent" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/edges/{edgeId}/status": {
      "get": {
        "tags": ["IoT Edge Gateway"],
        "summary": "Get Single Edge Gateway Status",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "edgeId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Edge agent details", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EdgeAgent" } } } },
          "404": { "description": "Edge agent not found" }
        }
      }
    },
...
    "/api/videos/upload": {
      "post": {
        "tags": ["Media & Creatives"],
        "summary": "Upload Video or Image Ad Creative",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["video"],
                "properties": {
                  "title": { "type": "string" },
                  "video": { "type": "string", "format": "binary" },
                  "thumbnail": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Media uploaded and processed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VideoMedia" } } } }
        }
      }
    },
    "/api/videos": {
      "get": {
        "tags": ["Media & Creatives"],
        "summary": "List User Creatives",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "List of creatives", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/VideoMedia" } } } } }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    },
    "schemas": {
...
      "EdgeAgent": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "edgeId": { "type": "string" },
          "siteId": { "type": "string" },
          "name": { "type": "string" },
          "agentVersion": { "type": "string" },
          "status": { "type": "string", "enum": ["online", "offline"] },
          "lastSeen": { "type": "string", "format": "date-time" },
          "telemetry": {
            "type": "object",
            "properties": {
              "cpuUsage": { "type": "number" },
              "memoryUsage": { "type": "number" },
              "activeStreams": { "type": "integer" }
            }
          }
        }
      },
...
      "VideoMedia": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "title": { "type": "string" },
          "url": { "type": "string" },
          "thumbnailUrl": { "type": "string" },
          "mediaType": { "type": "string" },
          "fileSize": { "type": "integer" }
        }
      }
    }
  }
}